20201210_update
The aim of this update was to give the user more freedom in selecting and merging user's input datasets.
Now there is no limit for the number of merges of user's files. Each merge always takes two files. => In the future I am planning to include brief note of the merging workflow history to give the user more information about the actions that have been done on the final dataset.
The last modification in this update includes warning messages about deleting the "output" folder and its content (the user is asked to confirm it each time whenever the deletion is going to be performed).